feat(scrap): - recurring button fully working - separate activebooking into smaller componentes
