+++ src/components/scrap-data/NewOperation/NewOperation.tsx
